+++ main.py
@@ -7,12 +7,16 @@ from dotenv import load_dotenv
 from functions import *
 
 
+# TODO
+# Add auto-leave on program exit
+# Detect channel change
+# Implement whitelist
+
+
 # Load environment
 load_dotenv()
 BOT_TOKEN = os.getenv('BOT_TOKEN')
 HOME_GUILD = os.getenv('HOME_GUILD')
-
-# TODO: Implement whitelist
 
 # Initialize bot
 intents = discord.Intents().none()
@@ -26,7 +30,6 @@ client = discord.Client(command_prefix='!', intents=intents)
 
 
 # Initialization behavior
-# TODO: Add auto-leave on exit
 @client.event
 async def on_ready():
     # Check for home guild (verbose output)
@@ -48,43 +51,55 @@ async def on_message(message):
     # Parse
     command = message.content.split(' ')[0]
     if not command[0] == '!': return
-    command = command[1:]
+    command = command[1:].lower()
 
     # Connect to voice channel
     if command == 'join':
+        await message.add_reaction('👍')
         await join(client, message)
 
     # Disconnect
     elif command == 'leave':
+        await message.add_reaction('👍')
         await disconnect(client, message)
 
     # Play music
     elif command == 'play':
+        await message.add_reaction('👍')
         await play(client, message)
 
     # Skip current song
     elif command == 'skip':
+        await message.add_reaction('👍')
         await skip(client, message)
 
     # Pause current song
     elif command == 'pause':
+        await message.add_reaction('👍')
         await pause(client, message)
 
     # Resume current song
     elif command == 'resume':
+        await message.add_reaction('👍')
         await resume(client, message)
 
     # Preview current song
     elif command == 'playing' or command == 'current':
+        await message.add_reaction('👍')
         await previewCurrent(client, message)
 
     # Preview next song
     elif command == 'next':
+        await message.add_reaction('👍')
         await previewNext(client, message)
 
     # Preview queue
     elif command == 'queue':
+        await message.add_reaction('👍')
         await previewQueue(client, message)
+
+    else:
+        await message.add_reaction('❌')
 
 # Run
 client.run(BOT_TOKEN)